requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
alembic>=1.13.0
python-multipart>=0.0.19
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0